Added new functions and moving to another brunch to change relationships

+++ src/main/java/com/khomsi/site_project/service/ProductService.java
@@ -4,6 +4,9 @@ import com.khomsi.site_project.entity.Product;
 import com.khomsi.site_project.exception.ProductNotFoundException;
 import com.khomsi.site_project.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,6 +14,15 @@ import java.util.NoSuchElementException;
 
 @Service
 public class ProductService implements IProductService {
+    public static final int PRODUCTS_PER_PAGE = 10;
+
+    public Page<Product> listByCategory(int pageNum, Integer categoryId) {
+        String categoryIdMatch = "-" + String.valueOf(categoryId) + "-";
+        Pageable pageable = PageRequest.of(pageNum - 1, PRODUCTS_PER_PAGE);
+
+        return productRepository.listByCategory(categoryId, pageable, categoryIdMatch);
+    }
+
     @Autowired
     private ProductRepository productRepository;
 
@@ -33,18 +45,21 @@ public class ProductService implements IProductService {
             throw new ProductNotFoundException("Couldn't find any product with id " + id);
         }
     }
-    //TODO реализовать поиск по тайтлу(алиасу)
-//    @Override
-//    public Product getProduct(String title) throws ProductNotFoundException {
-//        Product product = productRepository.findByCategory(title);
-//        if (product == null) {
-//            throw new ProductNotFoundException("Couldn't find any product with title " + title);
-//        }
-//        return product;
-//    }
+    @Override
+    public Product getProduct(String alias) throws ProductNotFoundException {
+        Product product = productRepository.findByAlias(alias);
+        if (product == null) {
+            throw new ProductNotFoundException("Couldn't find any product with alias " + alias);
+        }
+        return product;
+    }
 
     @Override
-    public void deleteProduct(int id) {
+    public void deleteProduct(Integer id) throws ProductNotFoundException {
+        Long countById = productRepository.countById(id);
+        if (countById == null || countById == 0){
+            throw new ProductNotFoundException("Couldn't find any product with ID " + id);
+        }
         productRepository.deleteById(id);
     }
 }
